my-shitty-git-operator/requirements.txt
kopf
kubernetes
requests
PyGithub